Refactor ModelPrior from model.util to model.extensions.
